backend/requirements-pip.txt
Django==3.1.3
django-extensions==3.0.9
django-3-jet===1.0.8
django-polymorphic==3.0.0
django-simple-history==2.12.0
djangorestframework==3.12.2
django-rest-knox==4.1.0
olefile==0.46
Pillow==8.0.1
pytz==2020.4
six==1.15.0
sorl-thumbnail==12.6.3
Werkzeug==1.0.1
django-taggit==1.3.0
psycopg2==2.8.6
django-crispy-forms==1.9.2
zeep==4.0.0
requests==2.25.0
sentry-sdk==0.19.4
django-cors-headers==3.5.0
django-dbbackup==3.3.0
django-storages==1.10.1
dropbox==10.10.0
